+++ lib/basf_flutter_components.dart
@@ -1,4 +1,8 @@
-export 'src/extensions/extensions.dart';
+export 'src/dto/text_field_data.dart';
+export 'src/extensions/list_extensions.dart';
+export 'src/extensions/object_extensions.dart';
+export 'src/extensions/other_extensions.dart';
+export 'src/extensions/string_extensions.dart';
 export 'src/theme/buttons.dart';
 export 'src/theme/colors.dart';
 export 'src/theme/dimens.dart';
@@ -17,6 +21,7 @@ export 'src/widgets/buttons/text_button.dart';
 export 'src/widgets/dialogs.dart';
 export 'src/widgets/inputs/logic/persisted_data.dart';
 export 'src/widgets/inputs/logic/persisted_input_cubit.dart';
+export 'src/widgets/inputs/logic/text_input_formatters.dart';
 export 'src/widgets/inputs/persisted_text_field.dart';
 export 'src/widgets/inputs/text_field.dart';
 export 'src/widgets/layout/label.dart';
